pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paramind"
version = "0.1.0"
description = "Dynamic Parallel Agentic Orchestration System"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.31.0",
    "aiohttp>=3.9.3",
    "openai>=1.12.0",
    "groq>=0.4.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
]

[tool.setuptools]
packages = ["src", "ui", "config"]